// src/api/generated/models/inventoryProductStockResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { InventoryProductStockResponseDtoStockState } from './inventoryProductStockResponseDtoStockState';
import type { InventoryStockVariantResponseDto } from './inventoryStockVariantResponseDto';

export interface InventoryProductStockResponseDto {
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  productId: string;
  productName: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  isSimple: boolean;
  variantCount: number;
  totalQuantity: number;
  stockState: InventoryProductStockResponseDtoStockState;
  updatedAt: string;
  variants: InventoryStockVariantResponseDto[];
}