// src/api/generated/models/stockReceiptListItemResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { StockReceiptListItemResponseDtoStatus } from './stockReceiptListItemResponseDtoStatus';
import type { StockReceiptPartyResponseDto } from './stockReceiptPartyResponseDto';
import type { StockReceiptUserResponseDto } from './stockReceiptUserResponseDto';

export interface StockReceiptListItemResponseDto {
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  id: string;
  code: string;
  /** @nullable */
  supplier?: StockReceiptPartyResponseDto | null;
  status: StockReceiptListItemResponseDtoStatus;
  itemCount: number;
  totalQuantity: number;
  totalCostAmount: string;
  /** @nullable */
  createdBy?: StockReceiptUserResponseDto | null;
  createdAt: string;
  /** @nullable */
  confirmedBy?: StockReceiptUserResponseDto | null;
  /** @nullable */
  confirmedAt?: string | null;
}